update default createAlbum to remove dedupe logic and create new getOrCreateAlbum for upload use case
